add ally opponent nearby calculation

+++ src/MinionStrategy/Expression.java
@@ -3,5 +3,5 @@ package MinionStrategy;
 import Game.Minion;
 
 public interface Expression extends Node{
-    public int evaluate(Minion minion) throws Exception;
+    public long evaluate(Minion minion) throws Exception;
 }
